requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.12
langchain-core==0.3.27
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.3
langsmith==0.1.143
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.14.1
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic-core==2.23.4
pydeck==0.9.1
pygments==2.18.0
pypdf2==3.0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.5.0
smmap==5.0.1
sniffio==1.3.1
sqlalchemy==2.0.36
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.2.2
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.2
openai==1.58.1
langchain-openai==0.2.13
distro==1.9.0
jiter==0.8.2
tenacity==9.0.0
tiktoken==0.8.0
dataclasses-json==0.6.7
httpx-sse==0.4.0
langchain-community==0.3.12
marshmallow==3.23.2
mypy-extensions==1.0.0
appdirs==1.4.4
datasets==3.2.0
ragas==0.2.8
dill==0.3.8
six==1.17.0
xxhash==3.5.0
pysbd==0.3.4
pyarrow==18.1.0
nest-asyncio==1.6.0
multiprocess==0.70.16
huggingface-hub==0.27.0
fsspec==2024.9.0
pypdfium2==4.30.1
pycparser==2.22
pdfplumber==0.11.4
pdfminer-six==20231228
cryptography==44.0.0
cffi==1.17.1